API/requirements.txt
alembic==1.5.2
asyncpg==0.21.0
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
databases==0.4.1
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
fastapi-users==5.0.0
h11==0.12.0
idna==3.1
makefun==1.9.5
Mako==1.1.4
MarkupSafe==1.1.1
passlib==1.7.4
psycopg2==2.8.6
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-multipart==0.0.5
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
uvicorn==0.13.3
